requirements.txt
fastapi
uvicorn[standard]
# psycopg2-binary is replaced by asyncpg
asyncpg
pgvector
torch
transformers
Pillow
numpy
retry
python-multipart
torchvision
